requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
psycopg2-binary>=2.9.7
chatmemory>=1.0.0
pyinstaller>=5.13.0
litellm>=1.0.0
